Create logical folders. Refactoring code. Implemented clicking on the list item
