+++ app/src/main/java/com/artyom_panfilenko/gwentdeckbuilder/DeckCard.java
@@ -1,13 +1,13 @@
 package com.artyom_panfilenko.gwentdeckbuilder;
 
 
-public class DeckCard extends Card{
+public class DeckCard extends Card {
 
     int num;
 
-    public DeckCard(int id,String name, String tags,String image,String description , String summon, String type, String rarity, String faction,int num) {
-    super(id,name,tags,image,description,type,rarity,faction);
-    this.num = num;
+    public DeckCard(int id, String name, String tags, String image, String description, String summon, String type, String rarity, String faction, int num) {
+        super(id, name, tags, image, description, type, rarity, faction);
+        this.num = num;
     }
 
     public int getNum() {
