+++ app/src/main/java/com/artyom_panfilenko/gwentdeckbuilder/MainActivity.java
@@ -3,17 +3,19 @@ package com.artyom_panfilenko.gwentdeckbuilder;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.DefaultItemAnimator;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.support.v7.widget.Toolbar;
-import android.view.View;
 import android.view.Menu;
 import android.view.MenuItem;
+import android.view.View;
 
-import java.util.ArrayList;
+import com.artyom_panfilenko.gwentdeckbuilder.activities.NewDeckActivity;
+import com.artyom_panfilenko.gwentdeckbuilder.adapters.DeckRVAdapter;
+import com.artyom_panfilenko.gwentdeckbuilder.database.DeckHelper;
+
 import java.util.List;
 
 public class MainActivity extends AppCompatActivity {
@@ -54,30 +56,31 @@ public class MainActivity extends AppCompatActivity {
 
     */
     RecyclerView rv;
-    List<Deck> decks = new ArrayList<>();
+    List<Deck> decks;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
+        Toolbar toolbar = findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
 
-        FloatingActionButton fab = (FloatingActionButton) findViewById(R.id.fab);
+        FloatingActionButton fab = findViewById(R.id.fab);
         fab.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Intent intent = new Intent(MainActivity.this,NewDeckActivity.class);
+                Intent intent = new Intent(MainActivity.this, NewDeckActivity.class);
                 startActivity(intent);
             }
         });
 
 
         DeckHelper deckHelper = new DeckHelper(this);
-        decks = deckHelper.getAll();
+        decks = deckHelper.getAllTest(10);
         rv = findViewById(R.id.rv);
         rv.setLayoutManager(new LinearLayoutManager(this));
         rv.setItemAnimator(new DefaultItemAnimator());
-        rv.setAdapter(new DeckRVAdapter(decks,this));
+        rv.setAdapter(new DeckRVAdapter(decks, this));
     }
 
     @Override
